Reintroduce a simpler overload process
Next to the manual writing of dispatching
overloads returning a specific callable object,
this commit reintroduce BOOST_DISPTACH_OVERLOAD
to take of simple overload cases with a macro.

No other macros will be provided to prevent feature
creep as we now have a simple manual way to specify
crooky cases.